refactor: split files into smaller pieces
